requirements.txt
redis==5.0.1
hiredis==2.3.2
pydantic==2.5.2
python-dotenv
setuptools
build
twine
